Created units templates Refined the supplier component, can now perform full search  and single selection search, TODO In supplier; lock and unlock fields depending on users operations

+++ src/app/suppliers/suppliers.component.ts
@@ -10,35 +10,38 @@ import { SupplierService } from '../supplier.service';
 })
 export class SuppliersComponent implements OnInit {
 
+  public suppliersNames: string [] = []
+  public suppliers = {}
+
   /**
    * field variables
    */
-  id
-  supplierCode
-  supplierName
-  contactName
-  tin
-  vrn
-  termsOfPayment
-  postAddress
-  postCode
-  physicalAddress
-  telephone
-  mobile
-  email
-  fax
-  bankAccountName
-  bankPostAddress
-  bankPostCode
-  bankName
-  bankAccountNo
-  bankStatus
+  id              : any
+  supplierCode    : string
+  supplierName    : string
+  contactName     : string
+  tin             : string
+  vrn             : string
+  termsOfPayment  : string
+  postAddress     : string
+  postCode        : string
+  physicalAddress : string
+  telephone       : string
+  mobile          : string
+  email           : string
+  fax             : string
+  bankAccountName : string
+  bankPostAddress : string
+  bankPostCode    : string
+  bankName        : string
+  bankAccountNo   : string
+  bankStatus      : string
 
   constructor(private httpClient: HttpClient ) {
     this.id               ='';
     this.supplierCode     ='';
     this.supplierName     ='';
-    this.contactName     ='';
+    this.contactName      ='';
     this.tin              ='';
     this.vrn              ='';
     this.termsOfPayment   ='';
@@ -57,9 +60,28 @@ export class SuppliersComponent implements OnInit {
     this.bankStatus       ='';
   }
 
+  async ngOnInit(): Promise<void> {
+    /**
+     * Load supplier names for autocompletion
+     * PRE:Empty array of suppliers names
+     * POS:A n array of supplier names
+     */
+    ((new SupplierService(this.httpClient)).getSuppliersNames())
+    .then(
+      res=>{
+        Object.values(res).map((supplierName:string)=>{
+          this.suppliersNames.push(supplierName)
+        })
+      }
+    );
+    this.suppliers = await (new SupplierService(this.httpClient)).getSuppliers()
+  }
+
   clear(){
     /**
      * clear the fields
+     * PRE:empty / non empty fields
+     * POS:empty fields
      */
     this.id               ='';
     this.supplierCode     ='';
@@ -86,6 +108,7 @@ export class SuppliersComponent implements OnInit {
   getSupplierData(){
     /**
      * gets item data from inputs
+     * RETURN supplier object
      */
     var supplierData = {
       id               :this.id,
@@ -140,19 +163,33 @@ export class SuppliersComponent implements OnInit {
   }
   validateData(){
     /**
-     * validates user inputs
-     * return false if validation fails
+     * Validates user inputs
+     * If validation fails,
+     * display error message and return false
+     * Otherwise, returns true
      */
     var valid : boolean = true
-    
-    
+    var message : string = ''
+    if(this.supplierCode == null || this.supplierCode == ''){
+      valid = false
+      message = "Supplier Code must not be empty!"
+      alert("Error: "+message)
+      return
+    }
+    if(this.supplierName == null || this.supplierName == ''){
+      valid = false
+      message = "Supplier Name must not be empty!"
+      alert("Error: "+message)
+      return
+    }
     return valid
   }
 
   saveSupplier() { 
     /**
-     * create or update an item
-     * first, check validation
+     * Create a new supplier, or update an existing supplier
+     * First validate user input
+     * 
      */
     if(this.validateData()==true){
       if( this.id == '' || this.id == null || this.id == 0 ) {
@@ -161,11 +198,11 @@ export class SuppliersComponent implements OnInit {
         .subscribe(
           data=>{
             window.alert('Supplier created successifully')
-            console.log(data)
             this.id=data['id']
           },
           error=>{
             console.log(error)
+            alert("Operation failed\nError Code:"+error['status']+"\n"+error['statusText'])
           }
         )
       } else {
@@ -174,11 +211,11 @@ export class SuppliersComponent implements OnInit {
         .subscribe(
           data=>{
             window.alert('Supplier updated successifully')
-            console.log(data)
             this.id=data['id']
           },
           error=>{
             console.log(error)
+            alert("Operation failed\nError Code:"+error['status']+"\n"+error['statusText'])
           }
         )
       }  
@@ -186,41 +223,64 @@ export class SuppliersComponent implements OnInit {
   }
   async searchSupplier() { 
     /**
-     * search supplier by id
-     * gets id from getSupplierId
+     * Search an existing supplier by supplier id
+     * First determine the supplier id using getSupplierId(), then searches
+     * supplier using respective supplier id
      */
-    var supplierId=0
-    this.clear
+    var supplierId = null
     supplierId = await (new SupplierService(this.httpClient)).getSupplierId(this.supplierCode , this.supplierName)
-    if(supplierId==0||supplierId==null){
-      alert('No matching record')
+    if(supplierId == null){
+      if(this.supplierCode == '' && this.supplierName == ''){
+        alert('Please enter a search key!')
+      }else{
+        alert('The requested record could not be found')
+      }
+      
     }else{
-      var supplier
-      supplier =await (new SupplierService(this.httpClient)).getSupplier(supplierId)
-      console.log(supplier)
+      var supplier =await (new SupplierService(this.httpClient)).getSupplier(supplierId)
       this.showSupplier(supplier)
     }
   }
+  async search(id : any){
+    this.clear()
+    var supplier =await (new SupplierService(this.httpClient)).getSupplier(id)
+    this.showSupplier(supplier)
+  }
   deleteSupplier() { 
     /**
-     * delete an item given its id
+     * Delete a supplier given supplier id
+     * PRE:Supplier record originally in the database
+     * POS:Supplier record removed from the database
      */
     var id = this.id
-    this.httpClient.delete(Data.baseUrl+"/suppliers/"+id)
-    .subscribe(
-      data=>{
-        console.log(data)
-        if(data==null){
-          this.clear()
-          alert('Supplier successifully deleted')
+    if(id == ''){
+      alert('Please select a supplier to delete')
+      return
+    }
+    if(window.confirm('Delete the selected supplier?\nThe supplier will be removed and this action can not be undone.\nConfirm?')){
+      this.httpClient.delete(Data.baseUrl+"/suppliers/"+id)
+      .subscribe(
+        data=>{
+          console.log(data)
+          if(data==null){
+            this.clear()
+            alert('Supplier Successiful deleted')
+          }
+        },
+        error=>{
+          alert('Could not delete the selected supplier')
         }
-      },
-      error=>{
-        console.log(error)
-      }
-    )
+      )
+    }
   }
-
-  ngOnInit(): void {}
-
+  //not used
+  lock : boolean = false
+  lockInputs(){
+    if(this.lock){
+      this.lock = false
+    }else if(!this.lock){
+      this.lock = true
+    }
+  }
+  
 }
